Adapt to the new datatype engine. There are 2 big differences with the previous version: 1. for the send operation: the convertor attached to the request is initialized by the    _INIT macro and it will be used as is by the PTLs for the first fragment. For the    others fragments it will be cloned by the PTLs. 2. for the receive operation: I create a macro MATCHED which create a convertor (attached    to the receive PML request). The PTLs can use this convertor for the first fragment,    and clone it forr the others fragments.
Playing with positions ... The position argument change it's type and prototype. Now it's
a size_t to be consistent with the rest of thee PML/PTL and it's a pointer to a size_t.
I need this pointer to correct the value of the position as some of the convertors are
not allowed to stop in the middle of a basic datatype (cannnot copy 2 bytes from an int).

This commit was SVN r5987.

+++ src/mca/pml/ob1/pml_ob1_sendreq.c
@@ -188,7 +188,7 @@ int mca_pml_ob1_send_request_start_copy(
 
         struct iovec iov;
         unsigned int iov_count;
-        unsigned int max_data;
+        size_t max_data;
         bool ack = false;
 
         /* determine first fragment size */
